XP-3033 Install Application dialog - Replace spinner with progress bar
- replaced spinner with api.ui.ProgressBar
- moved app loading to ApplicationService from ApplicationResource
- generated load progress events
- updated progress bar with values from these events

+++ modules/admin/admin-ui/src/main/resources/web/admin/common/js/application/ApplicationCache.ts
@@ -8,8 +8,10 @@ module api.application {
             super();
 
             ApplicationEvent.on((event: ApplicationEvent) => {
-                console.log("ApplicationCache on ApplicationEvent, deleting: " + event.getApplicationKey().toString());
-                this.deleteByKey(event.getApplicationKey());
+                if (event.getEventType() != ApplicationEventType.PROGRESS) {
+                    console.log("ApplicationCache on ApplicationEvent, deleting: " + event.getApplicationKey().toString());
+                    this.deleteByKey(event.getApplicationKey());
+                }
             });
         }
